[REFACTOR] Separate lock UI from lock slot logic
